requirement.txt
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1
PyYAML==6.0.1
numpy==1.24.4
timm==0.3.2
tensorboard==2.14.0
einops==0.7.0
pandas==2.0.3
matplotlib==3.7.4
tqdm==4.66.1
scipy==1.10.1
scikit-learn==1.3.2
ninja==1.11.1.1
packaging==23.2
h5py==3.9.0